Snapshot 76.0.3809.108

+++ src/storage/browser/fileapi/file_system_usage_cache.cc
@@ -24,7 +24,8 @@ constexpr base::TimeDelta kCloseDelay = base::TimeDelta::FromSeconds(5);
 const size_t kMaxHandleCacheSize = 2;
 }  // namespace
 
-FileSystemUsageCache::FileSystemUsageCache() : weak_factory_(this) {
+FileSystemUsageCache::FileSystemUsageCache(bool is_incognito)
+    : is_incognito_(is_incognito), weak_factory_(this) {
   DETACH_FROM_SEQUENCE(sequence_checker_);
 }
 
@@ -146,6 +147,8 @@ bool FileSystemUsageCache::UpdateUsage(const base::FilePath& usage_file_path,
 bool FileSystemUsageCache::Exists(const base::FilePath& usage_file_path) {
   TRACE_EVENT0("FileSystem", "UsageCache::Exists");
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (is_incognito_)
+    return base::ContainsKey(incognito_usages_, usage_file_path);
   return base::PathExists(usage_file_path);
 }
 
@@ -153,6 +156,12 @@ bool FileSystemUsageCache::Delete(const base::FilePath& usage_file_path) {
   TRACE_EVENT0("FileSystem", "UsageCache::Delete");
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
   CloseCacheFiles();
+  if (is_incognito_) {
+    if (!base::ContainsKey(incognito_usages_, usage_file_path))
+      return false;
+    incognito_usages_.erase(incognito_usages_.find(usage_file_path));
+    return true;
+  }
   return base::DeleteFile(usage_file_path, false);
 }
 
@@ -222,6 +231,10 @@ bool FileSystemUsageCache::Write(const base::FilePath& usage_file_path,
 
 base::File* FileSystemUsageCache::GetFile(const base::FilePath& file_path) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (is_incognito_) {
+    NOTREACHED();
+    return nullptr;
+  }
   if (cache_files_.size() >= kMaxHandleCacheSize)
     CloseCacheFiles();
   ScheduleCloseTimer();
@@ -248,6 +261,12 @@ bool FileSystemUsageCache::ReadBytes(const base::FilePath& file_path,
                                      char* buffer,
                                      int64_t buffer_size) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (is_incognito_) {
+    if (!base::ContainsKey(incognito_usages_, file_path))
+      return false;
+    memcpy(buffer, incognito_usages_[file_path].data(), buffer_size);
+    return true;
+  }
   base::File* file = GetFile(file_path);
   if (!file)
     return false;
@@ -258,6 +277,12 @@ bool FileSystemUsageCache::WriteBytes(const base::FilePath& file_path,
                                       const char* buffer,
                                       int64_t buffer_size) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (is_incognito_) {
+    if (!base::ContainsKey(incognito_usages_, file_path))
+      incognito_usages_[file_path] = std::vector<uint8_t>(buffer_size);
+    memcpy(incognito_usages_[file_path].data(), buffer, buffer_size);
+    return true;
+  }
   base::File* file = GetFile(file_path);
   if (!file)
     return false;
@@ -267,6 +292,8 @@ bool FileSystemUsageCache::WriteBytes(const base::FilePath& file_path,
 bool FileSystemUsageCache::FlushFile(const base::FilePath& file_path) {
   TRACE_EVENT0("FileSystem", "UsageCache::FlushFile");
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (is_incognito_)
+    return base::ContainsKey(incognito_usages_, file_path);
   base::File* file = GetFile(file_path);
   if (!file)
     return false;
