+++ src/testing/scripts/run_performance_tests.py
@@ -114,6 +114,16 @@ class OutputFilePaths(object):
   def logs(self):
     return os.path.join(self.benchmark_path, 'benchmark_log.txt')
 
+  @property
+  def csv_perf_results(self):
+    """Path for csv perf results.
+
+    Note that the chrome.perf waterfall uses the json histogram perf results
+    exclusively. csv_perf_results are implemented here in case a user script
+    passes --output-format=csv.
+    """
+    return os.path.join(self.benchmark_path, 'perf_results.csv')
+
 
 def print_duration(step, start):
   print 'Duration of %s: %d seconds' % (step, time.time() - start)
@@ -203,11 +213,15 @@ def execute_gtest_perf_test(command_generator, output_paths, use_xvfb=False):
   try:
     command = command_generator.generate()
     if use_xvfb:
+      # When running with xvfb, we currently output both to stdout and to the
+      # file. It would be better to only output to the file to keep the logs
+      # clean.
       return_code = xvfb.run_executable(
           command, env, stdoutfile=output_paths.logs)
     else:
-      return_code = test_env.run_command_with_output(
-          command, env=env, stdoutfile=output_paths.logs)
+      with open(output_paths.logs, 'w') as handle:
+        return_code = test_env.run_command_output_to_handle(
+            command, handle, env=env)
     # Get the correct json format from the stdout to write to the perf
     # results file.
     results_processor = generate_legacy_perf_dashboard_json.\
@@ -221,6 +235,11 @@ def execute_gtest_perf_test(command_generator, output_paths, use_xvfb=False):
     return_code = 1
   write_legacy_test_results(return_code, output_paths.test_results)
   return return_code
+
+
+class _TelemetryFilterArgument(object):
+  def __init__(self, filter_string):
+    self.benchmark, self.story = filter_string.split('/')
 
 
 class TelemetryCommandGenerator(object):
@@ -244,12 +263,13 @@ class TelemetryCommandGenerator(object):
     return ([sys.executable, self._options.executable] +
             [self.benchmark] +
             self._generate_filter_args() +
-            self._generate_repeat_args() +
             self._generate_also_run_disabled_tests_args() +
             self._generate_output_args(output_dir) +
             self._generate_story_range_args() +
-            # passthrough args must be before reference args: crbug.com/928928
+            # passthrough args must be before reference args and repeat args:
+            # crbug.com/928928, crbug.com/894254#c78
             self._get_passthrough_args() +
+            self._generate_repeat_args() +
             self._generate_reference_build_args()
            )
 
@@ -260,8 +280,11 @@ class TelemetryCommandGenerator(object):
     if self._options.isolated_script_test_filter:
       filter_list = common.extract_filter_list(
           self._options.isolated_script_test_filter)
+      filter_arguments = [_TelemetryFilterArgument(f) for f in filter_list]
+      applicable_stories = [
+          f.story for f in filter_arguments if f.benchmark == self.benchmark]
       # Need to convert this to a valid regex.
-      filter_regex = '(' + '|'.join(filter_list) + ')'
+      filter_regex = '(' + '|'.join(applicable_stories) + ')'
       return ['--story-filter=' + filter_regex]
     return []
 
@@ -297,8 +320,7 @@ class TelemetryCommandGenerator(object):
   def _generate_reference_build_args(self):
     if self._is_reference:
       return ['--browser=reference',
-              '--max-failures=5',
-              '--output-trace-tag=_ref']
+              '--max-failures=5']
     return []
 
 
@@ -318,15 +340,26 @@ def execute_telemetry_benchmark(
   try:
     command = command_generator.generate(temp_dir)
     if use_xvfb:
+      # When running with xvfb, we currently output both to stdout and to the
+      # file. It would be better to only output to the file to keep the logs
+      # clean.
       return_code = xvfb.run_executable(
           command, env=env, stdoutfile=output_paths.logs)
     else:
-      return_code = test_env.run_command_with_output(
-          command, env=env, stdoutfile=output_paths.logs)
+      with open(output_paths.logs, 'w') as handle:
+        return_code = test_env.run_command_output_to_handle(
+            command, handle, env=env)
+    expected_results_filename = os.path.join(temp_dir, 'test-results.json')
+    if os.path.exists(expected_results_filename):
+      shutil.move(expected_results_filename, output_paths.test_results)
+    else:
+      common.write_interrupted_test_results_to(output_paths.test_results, start)
     expected_perf_filename = os.path.join(temp_dir, 'histograms.json')
     shutil.move(expected_perf_filename, output_paths.perf_results)
-    expected_results_filename = os.path.join(temp_dir, 'test-results.json')
-    shutil.move(expected_results_filename, output_paths.test_results)
+
+    csv_file_path = os.path.join(temp_dir, 'results.csv')
+    if os.path.isfile(csv_file_path):
+      shutil.move(csv_file_path, output_paths.csv_perf_results)
   except Exception:
     print ('The following exception may have prevented the code from '
            'outputing structured test results and perf results output:')
@@ -394,8 +427,8 @@ def parse_arguments(args):
   return options
 
 
-def main():
-  args = sys.argv[1:]  # Skip program name.
+def main(sys_args):
+  args = sys_args[1:]  # Skip program name.
   options = parse_arguments(args)
   isolated_out_dir = os.path.dirname(options.isolated_script_test_output)
   overall_return_code = 0
@@ -405,6 +438,11 @@ def main():
   # since we do not monitor those. Also, merging test reference build results
   # with standard build results may not work properly.
   test_results_files = []
+
+  print('Running a series of performance test subprocesses. Logs, performance\n'
+        'results, and test results JSON will be saved in a subfolder of the\n'
+        'isolated output directory. Inside the hash marks in the following\n'
+        'lines is the name of the subfolder to find results in.\n')
 
   if options.non_telemetry:
     command_generator = GtestCommandGenerator(options)
@@ -416,6 +454,7 @@ def main():
     if not benchmark_name:
       benchmark_name = options.executable
     output_paths = OutputFilePaths(isolated_out_dir, benchmark_name).SetUp()
+    print('\n### {folder} ###'.format(folder=benchmark_name))
     overall_return_code = execute_gtest_perf_test(
         command_generator, output_paths, options.xvfb)
     test_results_files.append(output_paths.test_results)
@@ -428,6 +467,7 @@ def main():
         output_paths = OutputFilePaths(isolated_out_dir, benchmark).SetUp()
         command_generator = TelemetryCommandGenerator(
             benchmark, options)
+        print('\n### {folder} ###'.format(folder=benchmark))
         return_code = execute_telemetry_benchmark(
             command_generator, output_paths, options.xvfb)
         overall_return_code = return_code or overall_return_code
@@ -468,6 +508,7 @@ def main():
         output_paths = OutputFilePaths(isolated_out_dir, benchmark).SetUp()
         command_generator = TelemetryCommandGenerator(
             benchmark, options, stories=stories)
+        print('\n### {folder} ###'.format(folder=benchmark))
         return_code = execute_telemetry_benchmark(
             command_generator, output_paths, options.xvfb)
         overall_return_code = return_code or overall_return_code
@@ -479,6 +520,8 @@ def main():
           reference_command_generator = TelemetryCommandGenerator(
               benchmark, options,
               stories=stories, is_reference=True)
+          print('\n### {folder} ###'.format(
+              folder=reference_benchmark_foldername))
           # We intentionally ignore the return code and test results of the
           # reference build.
           execute_telemetry_benchmark(
@@ -490,8 +533,9 @@ def main():
 
   test_results_list = []
   for test_results_file in test_results_files:
-    with open(test_results_file, 'r') as fh:
-      test_results_list.append(json.load(fh))
+    if os.path.exists(test_results_file):
+      with open(test_results_file, 'r') as fh:
+        test_results_list.append(json.load(fh))
   merged_test_results = results_merger.merge_test_results(test_results_list)
   with open(options.isolated_script_test_output, 'w') as f:
     json.dump(merged_test_results, f)
@@ -513,4 +557,4 @@ if __name__ == '__main__':
       'compile_targets': main_compile_targets,
     }
     sys.exit(common.run_script(sys.argv[1:], funcs))
-  sys.exit(main())
+  sys.exit(main(sys.argv))
